rename testconfig package to jarvisconfig
