+++ src/main/java/ai/aliz/jarvis/jarvisconfig/ExecutionActionConfig.java
@@ -1,10 +1,8 @@
-package ai.aliz.jarvis.testconfig;
+package ai.aliz.jarvis.jarvisconfig;
 
 import lombok.Data;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 @Data
