+++ src/main/java/ai/aliz/jarvis/jarvisconfig/InitActionType.java
@@ -1,4 +1,4 @@
-package ai.aliz.jarvis.testconfig;
+package ai.aliz.jarvis.jarvisconfig;
 
 public enum InitActionType {
     BQLoad, SQLExec, SFTPLoad
