+++ src/main/java/ai/aliz/jarvis/service/assertaction/Assertor.java
@@ -1,6 +1,6 @@
 package ai.aliz.jarvis.service.assertaction;
 
-import ai.aliz.jarvis.testconfig.AssertActionConfig;
+import ai.aliz.jarvis.jarvisconfig.AssertActionConfig;
 
 public interface Assertor {
