+++ src/main/java/ai/aliz/jarvis/service/assertaction/MySQLAssertor.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 
 import ai.aliz.jarvis.context.TestContext;
 import ai.aliz.jarvis.context.TestContextLoader;
-import ai.aliz.jarvis.testconfig.AssertActionConfig;
+import ai.aliz.jarvis.jarvisconfig.AssertActionConfig;
 
 @Service
 public class MySQLAssertor implements Assertor {
