+++ src/main/java/ai/aliz/jarvis/service/initaction/initiator/Initiator.java
@@ -1,6 +1,6 @@
 package ai.aliz.jarvis.service.initaction.initiator;
 
-import ai.aliz.jarvis.testconfig.InitActionConfig;
+import ai.aliz.jarvis.jarvisconfig.InitActionConfig;
 
 public interface Initiator {
