+++ src/test/java/ai/aliz/jarvis/integration/TestInitActionServicePostgreSQLIntegration.java
@@ -18,9 +18,9 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import ai.aliz.jarvis.context.TestContextLoader;
-import ai.aliz.jarvis.testconfig.InitActionConfigFactory;
+import ai.aliz.jarvis.jarvisconfig.InitActionConfigFactory;
 import ai.aliz.jarvis.service.initaction.InitActionService;
-import ai.aliz.jarvis.testconfig.InitActionConfig;
+import ai.aliz.jarvis.jarvisconfig.InitActionConfig;
 
 import org.junit.BeforeClass;
 import org.junit.Rule;
